[Xamarin.Andorid.Build.Tasks] First Pass on Dynamic Asset Features
Context #4810

This is the first implementation for supporting building Dynamic Feature
Assets modules for Android.  This idea is to have these "features" as
standard Xamarin Android Library projects. The `ProjectReference` will need to
use the following

```xml
<ItemGroup>
    <ProjectReference Include="Features\AssetsFeature\AssetsFeature.csproj">
         <ReferenceOutputAssembly>false</ReferenceOutputAssembly>
         <AndroidDynamicFeature>true</AndroidDynamicFeature>
    </ProjectReference>
  </ItemGroup>
```

We set `ReferenceOutputAssembly` to `false` to ensure that the assembly
for the feature is NOT included in the final aab file.  The new meta
Data `AndroidDynamicFeature` allows the build system to pick up project
references which are "features".

As part of the final packaging step of the main app we will gather up all
the `ProjectRefernce` items which have `AndroidDynamicFeature` set to
`true` (and maybe `ReferenceOutputAssembly` set to `false`). This will
be done by the `_BuildDynamicFeatures` which will run just after
`_CreateBaseApk`.

It will call `_GetDynamicFeatureOutputs` for each `ProjectReference`
which will collect the `output` files for each feature. It will then call
the `BuildDynamicFeature` target via the `MSBuild` task for each
`ProjectReference`.  The `BuildDynamicFeature` is the target responsible
for collecting all the assets and packaging them using `aapt2` up into a
zip. Once all the `BuildDynamicFeature` calls are complete the created
`zip` files will be added to the `AndroidAppBundleModules` and then
included in the final `aab` file.

It might seem odd that the feature projects are built after the main app.
However this is required because the  feature needs to use the
`packaged_resources` file as an input to `aapt2` when building the
feature. This is why the `_BuildDynamicFeatures` happens AFTER `_CreateBaseApk`.
It is only at that point that the final `packaged_resources` file exists.

One of the very weird things is that the feature zip needs to be built
using the `aapt2` `--static-lib` flag. As a result we need to call
`aapt2 convert` on the final zip. This is because it is in the `apk`
`binary` format and needs to be converted over to the `aab` `proto` format.
So there is a new `Aapt2Convert` task which handles that job. It also
makes sure the `AndroidManifest.xml` file is in the right place when
converting to `proto` format.

A basic project example using .net 6 for a feature would look like this.

```xml
<Project Sdk="Microsoft.NET.Sdk">
  <PropertyGroup>
    <TargetFramework>net6.0-android</TargetFramework>
  </PropertyGroup>
</Project>
```

As you can see it is just a normal library project. At this time is CANNOT
contain any `AndroidResource` items such as drawables or layouts. It must
only contain `AndroidAsset` items. So we probably should have a new
template for a `Dynamic Feature` which just creates the `csproj` and
the `Assets` folder.

One sticking point is probably the `AndroidManifest.xml` file which we
need for a `feature`. There is a sample

```xml
<?xml version="1.0" encoding="utf-8"?>
<manifest xmlns:android="http://schemas.android.com/apk/res/android"
        xmlns:tools="http://schemas.android.com/tools"
        xmlns:dist="http://schemas.android.com/apk/distribution"
        android:versionCode="1"
        android:versionName="1.0"
        package="com.companyname.DynamicAssetsExample"
        featureSplit="assetsfeature"
        android:isFeatureSplit="true">
  <dist:module dist:title="@string/assetsfeature" dist:instant="false">
    <dist:delivery>
      <dist:on-demand />
    </dist:delivery>
    <dist:fusing dist:include="false" />
  </dist:module>
  <uses-sdk android:minSdkVersion="21" android:targetSdkVersion="30" />
  <application android:hasCode="false" tools:replace="android:hasCode" />
</manifest>
```

The interesting parts are all the additional `dist` elements. What we can
probably do is auto generate this during the `BuildDynamicFeature`.
However we need to think carefully about this since if we plan to have
code and `Activities` in the feature at some point, those will also need
to end up in the `AndroidManifest.xml`.

For additional information on the Play Core Dynamic Features check the
following links.
[1] https://developer.android.com/guide/playcore/asset-delivery
[2] https://developer.android.com/guide/playcore/feature-delivery