[Xamarin.Android.Build.Tasks] require JDK 11 for API 31
API 31 builds fail when using JDK 1.8 with:

    /Library/Frameworks/Mono.framework/External/xbuild/Xamarin/Android/Xamarin.Android.Common.targets(1728,3):
    error JAVAC0000: java.lang.AssertionError: annotationType() : unrecognized Attribute name MODULE (class com.sun.tools.javac.util.SharedNameTable$NameImpl)

To solve this issue, we must require JDK 11 for API 31.

There are two code paths here for "legacy" Xamarin.Android and .NET 6.
Legacy is straightforward, as we already had a pattern for this.

In .NET 6, I checked the `$(TargetPlatformVersion)` property for now
and require JDK 11 in that case. When API 31 is stable, we can
probably remove this code and just set this value by default:

    <MinimumSupportedJavaVersion Condition=" '$(MinimumSupportedJavaVersion)' == '' ">11.0</MinimumSupportedJavaVersion>

* Add new `AssertTargetFrameworkVersionSupported()` method

In tests where it makes sense, I used a different
`$(TargetFrameworkVersion)` or just removed what was setting the
property entirely.

Other tests, I used a `AssertTargetFrameworkVersionSupported()` method
to `Assert.Ignore()` the test if it would not pass otherwise.